get_binance_book added, foundation for signed endpoint security added
